swarm: Move tests to a separate file
